added test for pathfinding and refactored file locations
