+++ src/main/java/frc/robot/Constants.java
@@ -3,12 +3,17 @@ package frc.robot;
 import static edu.wpi.first.units.Units.Inches;
 import static edu.wpi.first.units.Units.MetersPerSecond;
 
+import com.pathplanner.lib.path.PathConstraints;
+
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Rotation3d;
 import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.LinearVelocity;
 import edu.wpi.first.wpilibj.RobotBase;
+import frc.robot.generated.TunerConstants;
 
 public class Constants {
   /**
@@ -133,5 +138,36 @@ public class Constants {
     public static final double FF_RAMP_RATE = 0.1; // Volts/Sec
     public static final double WHEEL_RADIUS_MAX_VELOCITY = 0.25; // Rad/Sec
     public static final double WHEEL_RADIUS_RAMP_RATE = 0.05; // Rad/Sec^2
+  }
+
+  public final class PathGenerationConstants {
+
+
+    // Predefined locations of interest
+    public enum Location {
+      SPEAKER_CENTER(new Pose2d(0.0, 5.5, Rotation2d.fromDegrees(0))), // Example coordinates
+      AMP(new Pose2d(1.8, 7.7, Rotation2d.fromDegrees(90))),
+      SOURCE_RIGHT(new Pose2d(15.5, 1.0, Rotation2d.fromDegrees(120))),
+      STAGE_CENTER(new Pose2d(4.5, 4.0, Rotation2d.fromDegrees(0)));
+
+      private final Pose2d pose;
+
+      Location(Pose2d pose) {
+        this.pose = pose;
+      }
+
+      public Pose2d getPose() {
+        return pose;
+      }
+    }
+   
+  // Default constraints for pathfinding
+    // Adjust these based on your robot's capabilities
+    public static final PathConstraints DEFAULT_CONSTRAINTS =
+        new PathConstraints(
+            TunerConstants.kSpeedAt12Volts.in(MetersPerSecond), // Max velocity (m/s)
+            7.9, // Max acceleration (m/s^2) (from PathPlanner)
+            TunerConstants.kSpeedAt12Volts.in(MetersPerSecond) / Math.sqrt(Math.pow(0.273, 2) * 2), // Max angular velocity (rad/s) calculated by dividing velocity by wheel radius
+            Units.degreesToRadians(790)); // Max angular acceleration (rad/s^2) (from PathPlanner)
   }
 }
