+++ src/main/java/frc/robot/util/PathGeneration.java
@@ -1,60 +1,21 @@
-package frc.robot.subsystems.pathgeneration;
+package frc.robot.util;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.path.PathConstraints;
 import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.Constants.PathGenerationConstants;
+import frc.robot.Constants.PathGenerationConstants.Location;
+
 import java.util.HashMap;
 import java.util.Map;
 
-import org.littletonrobotics.junction.Logger;
-
-public class PathGeneration extends SubsystemBase {
-  private final PathGenerationIO io;
-  private final PathGenerationIOInputsAutoLogged inputs = new PathGenerationIOInputsAutoLogged();
-
-  // Default constraints for pathfinding
-  // Adjust these based on your robot's capabilities
-  private static final PathConstraints DEFAULT_CONSTRAINTS =
-      new PathConstraints(
-          3.0, // Max velocity (m/s)
-          3.0, // Max acceleration (m/s^2)
-          Units.degreesToRadians(540), // Max angular velocity (rad/s)
-          Units.degreesToRadians(720)); // Max angular acceleration (rad/s^2)
-
-  // Predefined locations of interest
-  public enum Location {
-    SPEAKER_CENTER(new Pose2d(0.0, 5.5, Rotation2d.fromDegrees(0))), // Example coordinates
-    AMP(new Pose2d(1.8, 7.7, Rotation2d.fromDegrees(90))),
-    SOURCE_RIGHT(new Pose2d(15.5, 1.0, Rotation2d.fromDegrees(120))),
-    STAGE_CENTER(new Pose2d(4.5, 4.0, Rotation2d.fromDegrees(0)));
-
-    private final Pose2d pose;
-
-    Location(Pose2d pose) {
-      this.pose = pose;
-    }
-
-    public Pose2d getPose() {
-      return pose;
-    }
-  }
+public class PathGeneration {
 
   private final Map<String, Pose2d> customLocations = new HashMap<>();
 
-  public PathGeneration(PathGenerationIO io) {
-    this.io = io;
-  }
-
-  @Override
-  public void periodic() {
-    io.updateInputs(inputs);
-    Logger.processInputs("PathGeneration", inputs);
-  }
+  public PathGeneration() {}
 
   /**
    * Generates a command to pathfind to a specific known location.
@@ -63,7 +24,7 @@ public class PathGeneration extends SubsystemBase {
    * @return The pathfinding command.
    */
   public Command pathfindTo(Location location) {
-    return pathfindTo(location.getPose(), DEFAULT_CONSTRAINTS);
+    return pathfindTo(location.getPose(), PathGenerationConstants.DEFAULT_CONSTRAINTS);
   }
 
   /**
@@ -84,7 +45,7 @@ public class PathGeneration extends SubsystemBase {
    * @return The pathfinding command using default constraints.
    */
   public Command pathfindTo(Pose2d targetPose) {
-    return pathfindTo(targetPose, DEFAULT_CONSTRAINTS);
+    return pathfindTo(targetPose, PathGenerationConstants.DEFAULT_CONSTRAINTS);
   }
 
   /**
@@ -108,7 +69,7 @@ public class PathGeneration extends SubsystemBase {
    * @return A ParallelDeadlineGroup containing the pathfinding command and the parallel command.
    */
   public Command pathfindToWithHook(Location location, Command parallelCommand) {
-    return pathfindToWithHook(location.getPose(), DEFAULT_CONSTRAINTS, parallelCommand);
+    return pathfindToWithHook(location.getPose(), PathGenerationConstants.DEFAULT_CONSTRAINTS, parallelCommand);
   }
 
   /**
